requirements.txt
fastapi
uvicorn
pymongo
gspread
pydantic
python-dotenv
